requirements.txt
blinker==1.6.3
boto3==1.28.74
botocore==1.31.74
click==8.1.7
coloredlogs==15.0.1
Flask==3.0.0
Flask-Cors==4.0.0
flatbuffers==23.5.26
humanfriendly==10.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
logging==0.4.9.6
MarkupSafe==2.1.3
mpmath==1.3.0
numpy==1.26.1
onnxruntime==1.16.1
opencv-python==4.8.1.78
packaging==23.2
Pillow==10.1.0
protobuf==4.24.4
python-dateutil==2.8.2
s3transfer==0.7.0
six==1.16.0
sympy==1.12
urllib3==2.0.7
Werkzeug==3.0.1
